+++ src/Symfony/Component/Form/Extension/Core/Type/IntegerType.php
@@ -25,7 +25,7 @@ class IntegerType extends AbstractType
     {
         $builder->addViewTransformer(
             new IntegerToLocalizedStringTransformer(
-                $options['scale'],
+                null,
                 $options['grouping'],
                 $options['rounding_mode']
         ));
@@ -37,8 +37,6 @@ class IntegerType extends AbstractType
     public function configureOptions(OptionsResolver $resolver)
     {
         $resolver->setDefaults(array(
-            // default scale is locale specific (usually around 3)
-            'scale' => null,
             'grouping' => false,
             // Integer cast rounds towards 0, so do the same when displaying fractions
             'rounding_mode' => IntegerToLocalizedStringTransformer::ROUND_DOWN,
@@ -55,7 +53,9 @@ class IntegerType extends AbstractType
             IntegerToLocalizedStringTransformer::ROUND_CEILING,
         ));
 
+        $resolver->setDefined('scale');
         $resolver->setAllowedTypes('scale', array('null', 'int'));
+        $resolver->setDeprecated('scale');
     }
 
     /**